Рефакторинг проведен. Коммы добавлены. Root namespace акутализирован.
