--- Assets/IDestroyable.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-public interface IDestroyable
-{
-    bool IsDestroyed { get; }
-    
-    event Action<IDestroyable> OnObjectDestroyed;
-
-    void Destroy();
-}
