--- Assets/IHitable.cs
@@ -1,6 +1,0 @@
-using FlavorfulStory.Actions;
-
-public interface IHitable
-{
-    void TakeHit(ToolType toolType);
-}
