+++ Assets/Scripts/Actions/Interactables/AbstractHarvestableObject.cs.meta
