+++ Assets/Scripts/Actions/Interactables/HarvestableOnce.cs.meta
