+++ Assets/Scripts/Actions/Interactables/HarvestableWithCooldown.cs.meta
