+++ Assets/Scripts/Actions/Tool.cs
@@ -1,23 +1,25 @@
 using FlavorfulStory.Control;
 using FlavorfulStory.InventorySystem;
+using FlavorfulStory.ResourceContainer;
 using UnityEngine;
 
 namespace FlavorfulStory.Actions
 {
     /// <summary> Инструмент, используемый игроком для взаимодействия с объектами. </summary>
     /// <remarks> Может выполнять действия, специфичные для типа инструмента. </remarks>
-    [CreateAssetMenu(menuName = ("FlavorfulStory/Inventory/Tool"))]
+    [CreateAssetMenu(menuName = "FlavorfulStory/Inventory/Tool")]
     public class Tool : ActionItem
     {
-        /// <summary> Тип инструмента. </summary>
-        [field: Tooltip("Тип инструмента."), SerializeField]
-        public ToolType ToolType { get; private set; }
-
         /// <summary> Максимальная дистанция взаимодействия. </summary>
         private const float MaxInteractionDistance = 2f;
 
         /// <summary> Радиус использования инструмента. </summary>
         private const float UseRadius = 1.5f;
+
+        /// <summary> Тип инструмента. </summary>
+        [field: Tooltip("Тип инструмента.")]
+        [field: SerializeField]
+        public ToolType ToolType { get; private set; }
 
         /// <summary> Использовать инструмент. </summary>
         /// <param name="player"> Контроллер игрока. </param>
@@ -41,12 +43,10 @@ namespace FlavorfulStory.Actions
             var direction = (targetPosition - origin).normalized;
             var interactionCenter = origin + direction * (MaxInteractionDistance / 2);
             var hitColliders = Physics.OverlapSphere(interactionCenter, UseRadius);
-            
+
             foreach (var collider in hitColliders)
-            {
                 if (collider.TryGetComponent<IHitable>(out var hitable))
                     hitable.TakeHit(ToolType);
-            }
 
             Debug.DrawLine(origin, interactionCenter, Color.red, 5f);
         }
