+++ Assets/Scripts/Actions/ToolPrefabMapping.cs
@@ -1,17 +1,16 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace FlavorfulStory.Actions
 {
     /// <summary> Сопоставление типа инструмента с его префабом. </summary>
-    [System.Serializable]
+    [Serializable]
     public class ToolPrefabMapping
     {
         /// <summary> Тип инструмента. </summary>
-        [Tooltip("Тип инструмента.")]
-        public ToolType ToolType;
+        [Tooltip("Тип инструмента.")] public ToolType ToolType;
 
         /// <summary> Префаб инструмента. </summary>
-        [Tooltip("Префаб инструмента.")]
-        public GameObject ToolPrefab;
+        [Tooltip("Префаб инструмента.")] public GameObject ToolPrefab;
     }
 }
