+++ Assets/Scripts/ObjectManagement/DestroyableResourceContainerSpawner.cs.meta
