+++ Assets/Scripts/ObjectManagement/ObjectSwitcher.cs
@@ -1,0 +1,50 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace FlavorfulStory.ObjectManagement
+{
+    public class ObjectSwitcher : MonoBehaviour
+    {
+        /// <summary> Префабы грейдов одного объекта. </summary>
+        /// <remarks> Версия по умолчанию(первый ребенок в иерархии) добавляется автоматически. </remarks>
+        [Tooltip("Префабы всех грейдов данного объекта, кроме начального")] [SerializeField]
+        private GameObject[] _objectPrefabs;
+
+        /// <summary> Объекты на сцене, представляющие грейды разного уровня. </summary>
+        private List<GameObject> _spawnedObjects;
+
+        /// <summary> Инициализировать все грейды. </summary>
+        /// <remarks>
+        ///     В качестве первого грейда выступает дочерний объект под индексом 0.
+        ///     Многократный вызов не приведет к ошибке
+        /// </remarks>
+        public void Initialize()
+        {
+            if (_spawnedObjects != null) return;
+
+            _spawnedObjects = new List<GameObject>(_objectPrefabs.Length + 1) { transform.GetChild(0).gameObject };
+            foreach (var prefab in _objectPrefabs)
+            {
+                var spawnedObject = Instantiate(prefab, transform, false);
+                spawnedObject.SetActive(false);
+                _spawnedObjects.Add(spawnedObject);
+            }
+        }
+
+        /// <summary> Переключится на выбранный грейд. </summary>
+        /// <remarks> Включает одного из детей данного Gameobject'а. Выключает остальные. </remarks>
+        /// <param name="index"> Индекс грейда. </param>
+        public void SwitchTo(int index)
+        {
+            for (var i = 0; i < _spawnedObjects.Count; i++)
+                _spawnedObjects[i].SetActive(i == index);
+        }
+
+        /// <summary> Получить количество грейдов. </summary>
+        /// <returns> Количество грейдов в свитчере. </returns>
+        public int GetObjectsCount()
+        {
+            return _objectPrefabs.Length + 1;
+        }
+    }
+}
