+++ Assets/Scripts/ResourceContainer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0304f01784ff4bd4590c75c7a203298d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
